fix(lwd): reorganise the file structure and namings
