+++ apps/ledger-live-desktop/src/mvvm/features/GenericAwarenessModal/hooks/useGenericAwarenessModalViewModel.test.ts
@@ -1,6 +1,9 @@
 import { act, renderHook } from "tests/testSetup";
 import type { AppDispatch } from "~/state-manager/configureStore";
-import { closeGenericAwarenessModalDialog, openGenericAwarenessModalDialog } from "./genericAwarenessModalDialog";
+import {
+  closeGenericAwarenessModalDialog,
+  openGenericAwarenessModalDialog,
+} from "../genericAwarenessModalDialog";
 import useGenericAwarenessModalViewModel from "./useGenericAwarenessModalViewModel";
 
 const dispatchThunk = (
