+++ apps/ledger-live-desktop/src/mvvm/features/GenericAwarenessModal/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import GenericAwarenessModalView from "./GenericAwarenessModalView";
-import useGenericAwarenessModalViewModel from "./useGenericAwarenessModalViewModel";
+import useGenericAwarenessModalViewModel from "./hooks/useGenericAwarenessModalViewModel";
 
 const GenericAwarenessModal = () => (
   <GenericAwarenessModalView {...useGenericAwarenessModalViewModel()} />
